Instantiator: Make asynchronous mode work
This patch proceeds in two steps. First, and the most trivial,
we make sure not to offend the Instantiator's objects container
since asynchronous creation also means the objects can be created
in any order. We do so by ensuring the objects container has
always the necessary size.

The second step is to make sure the objects don't get destroyed by
the delegate model when their incubation tasks are asynchronous.
We force to keep an extra reference calling the object() function
on the instance model, but only if the creation was asynchronous.
However, it's not enough to check for the Instantiatior async value
since the incubation tasks in QQmlDelegateModel are async-if-nested.
Therefore, we need to keep track of the object index requested to
the model and see if it matches when the createdItem() signal gets
emitted.

Task-number: QTBUG-36410
Change-Id: I180b03b6a7468d6521827a9631755577e9e6d357
Reviewed-by: Andrew den Exter <andrew.den.exter@qinetic.com.au>