Button: provide a default size for the label delegate
If the label delegate has no explicit size defined, make it follow the
size of the button minus padding. This should be a convenient default
behavior that satisfies the most users, and saves a few of QML bindings.
Makes also the upcoming "the default delegate implementation looks like
this" -snippet nice and short. ;)

Change-Id: I574439c1e52c2904746be7d830f9536d4cfd9083
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>