ControlSpy: Accept a list of signals to spy on
This API is cleaner than having to connect to each signal
handler namually.

We solve the 'pressed' signal problem by connecting to the
other signal accessor, namely 'onPressed'. And we do so for
all signals.

We also lay foundation to capture the signal parameters as
each signal is emitted. What we may do with those values is
not clear yet.

Auto-tests updated accordingly.

Change-Id: Ieda96a43072ca7dff2b183e07f6f67aa49c8cf5d
Reviewed-by: Mitch Curtis <mitch.curtis@digia.com>