SignalSpy: Support "pressed" signal use-case
In some types, e.g. MouseArea, the "pressed" symbol refers
both to a signal and to a property. When trying to connect
to the signal, we get an error since the engine will resolve
to the property instead. We circumvent that by connecting to
the signal handler instead, i.e. "onPressed" for the "pressed"
signal.

Change-Id: I4212b752be4da4ec2209a3447ac41326f1e0bb5d
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>