ApplicationWindow: Add interstitial content item
The goal of this is twofold. First, it allows the window's
contentItem to follow the window size, keeping its nice
properties. And second, it prepares for adding a popup
layer item to ApplicationWindow, which should remove any
need to set high z values to ensure popup windows stay on top.

Change-Id: Ie024b7627808e2f3d22d7572f0ac5975e9bdf1a5
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>