add QQuickPointerEvent and QQuickPointerDevice
QQuickPointerEvent is first of all a wrapper for mouse and touch events
to help unify event delivery.  Because the instances are reused between
deliveries (as in change 0ac9389dbd7facaa424dc175e5f5dac514630eef),
it is allowed to be a QObject.  This will facilitate exposing
pointer events to QML later on, as with the other event classes
in qquickevents_p_p.h.  It allows both device-agnostic and
device-specific handling of point-based interaction using mice
and touchscreens (tablet support will come later).

The interface of QQuickPointerDevice is a superset of QTouchDevice
plus the relevant mouse and tablet device info.  The instances are
long-lived, and included by pointer in each QQuickPointerEvent.
Applications and device-specific event handlers can use the device
info to filter out irrelevant events.

Change-Id: Iacb42d41889a9a8634d57331ec6cc65c6b36f71f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>