Don't accept left clicks when text format is plain.
Text elements may contain rich text with embedded links, and
need to accept left clicks to open them. However, setting the
textFormat to PlainText can disable mouse handling entirely,
as it is not required.

Accepting left clicks if there can be nothing to interact with
is unexpected and surprising, and can cause bugs in code that
performs child event filtering and doesn't expect Text elements
to produce child events.

Change-Id: Ibd5b9cf8d06fd30ea26f78b5393cc43e94646e73
Reviewed-by: Marco Martin <mart@kde.org>
Reviewed-by: Kai Uwe Broulik <kde@privat.broulik.de>
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>