Remove TabBar::highlight
It could have been a convenient way to customize the default
highlight, but let's not expose an alias to a replaceable content
item. The alias might become invalid and meaningless, depending on
what kind of content item the replacement is. Thus, remove the
content item identifier to avoid the pitfall in the future.

Change-Id: I65d66e21b57a21a87cba03c5f5825d18092d0c39
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>