Refactor ExclusiveGroup
Add support for toggled() and toggled(bool) signals to bring
ExclusiveGroup 2.0 on par with ExclusiveGroup 1.0. Furthermore,
the attached property is now "ExclusiveGroup.group" to be consistent
with all other attached properties like StackView.view, TabView.view,
and others. It looks clumsy to repeat, but it's better to keep the
type names to the minimum, and the documentation also stays in one
place.

Change-Id: Ic46f8d2ec4b58c5056e8b2bbca5c9f88968bfb80
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>