Adding more DesignerSupport functions
The new functions/enablers are split into new classes.

* DesignerSupportItems
  Helper functions to tweak and create items.

* DesignerSupportProperties
  Helper functions to deal with properties and bindings.

* DesignerSupportPropertyChanges
  Helper functions to deal with PropertyChanges.

* DesignerSupportStates
  Helper functions to deal with States.

* DesignerSupportMetaInfo
  Two helper functions to deal with meta info and notication.

DesignerMetaObject is a specialized QQmlVMEMetaObject that implments
notification of property changes and allows to create dynamic properties.

DesignerCustomObjectData provides additional data for each Object. The
main functionality atm is to keep the default values and bindings for
properties, so they can be restored.

Change-Id: I6b6f501c8c5848d62bfcb7e704bc45662e731d54
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@digia.com>