Introducing Panel type
This is the base type to implement all sorts of Popups and Dialogs.
Its main role is to keep the panel contents stacked on top of the
application window contents and to ensure mouse and key events are
forwarded or blocked depending on the visible panels modality.

Currently, it only works with ApplicationWindow, which holds a
QQuickOverlay. This special item is where the Panel contents gets
reparented when the Panel becomes visible. It's also responsible
for filtering the mouse events.

Future developements may include adding a 'level' property instead
of relying on the item's z value. This may or may not result into
having several overlays per window.

Change-Id: I18a4b8905e4d5a4a4697642b0553a1f9e86b669f
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>