Emit movement signals for flick().
Make flick() more like a real flick and ensure the movement signals
and properties are updated. This allows them to be handled from QML.

This also fixes issues with flick() and dynamic delegates. Flickable
has several checks of the form:

    !d->pressed && !d->hData.moving && !d->vData.moving

That were processed incorrectly for flick(), as the moving variables
were not being updated.

[ChangeLog][QtQuick][Flickable] The movement related signals and
properties are now updated for flicks started via the flick function.

Change-Id: I7e96e2e12a4d0a0ee73ddd6f29d95f19c44667b0
Task-number: QTBUG-34507
Reviewed-by: Martin Jones <martin.jones@qinetic.com.au>