Add gifs manual test.
This is similar to the snippets auto test, in that it's more of a
utility than a test. In this case, the testing provides a convenient
way to run several recording "jobs" at once, as well as feed the app
with generated input events.

The test is different to the snippets auto test in that it:

- Tests QML code that isn't shown to users; only the GIFs are displayed
  in the documentation.
- Requires interaction to produce its output, which is achieved with
  mousePress(), etc. For this reason, it also can't just iterate over
  the list of QML files.
- Requires byzanz to be installed in order to record the GIFs.

The test also comes with EventCapturer, an event-filter based
class that records mouse movements and generates compact C++ code for
use with the GifRecorder.

There is one known issue, which is a bug in byzanz-record: it sometimes
decides to record the mouse cursor even if you didn't ask it to.

Change-Id: Icaaa4f37c8d34a813e36901fd187d84e4f250d33
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>