QQmlListModel: Optimize model updates
Only update changed indices in the list model on insertion, deletion or
move. This can greatly improves performance when inserting or removing
near the end of the list, or when moving elements in a large model.

On a work in progress version of qmlbench's modelperf_listmodel_addremove.qml,
this appears to improve performance quite a bit (~40%):

Before: Average: 363.4  frames; using  samples; MedianAll=363; StdDev=5.31977, CoV=0.0146389
 After: Average: 518.2  frames; using  samples; MedianAll=518; StdDev=20.5475, CoV=0.0396517

Change-Id: I8ce7bea4ab8b1c7c3d0fbccc40b9a12ca3eadf59
Reviewed-by: Robin Burchell <robin.burchell@crimson.no>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>